+++ source/App/source/ExampleHost.WebApi02/Startup.cs
@@ -29,12 +29,7 @@ public class Startup
     {
         services.AddControllers();
 
-        // Configuration verified in tests
-        //  * Logging using ILogger<T> will work, but notice that by default we need to log as "Warning" for it to appear in Application Insights (can be configured).
-        //    See "How do I customize ILogger logs collection" at https://learn.microsoft.com/en-us/azure/azure-monitor/app/asp-net-core?tabs=netcorenew#how-do-i-customize-ilogger-logs-collection
-        //  * We can see Trace, Request, Dependencies and other entries in App Insights out-of-box.
-        //    See https://docs.microsoft.com/en-us/azure/azure-monitor/app/asp-net-core
-        //  * Telemetry events are enriched with property "Subsystem" and configured value
+        // Configuration verified in tests. See comments in WebApi01.Startup.
         services.AddApplicationInsightsForWebApp(subsystemName: "ExampleHost.WebApp");
     }